update migrate for stg

+++ database/migrations/2021_08_13_114818_add_rank_to_cpa_table.php
@@ -26,7 +26,8 @@ class AddRankToCpaTable extends Migration
     public function down()
     {
         Schema::table('cpa', function (Blueprint $table) {
-            $table->dropColumn(['rank']);
+            if (Schema::hasColumn('cpa', 'rank'))
+                $table->dropColumn(['rank']);
         });
     }
 }
