+++ database/migrations/2021_08_17_070145_add_program_id_to_cpa_table.php
@@ -28,7 +28,8 @@ class AddProgramIdToCpaTable extends Migration
     public function down()
     {
         Schema::table('cpa', function (Blueprint $table) {
-            $table->dropColumn('program_id');
+            if (Schema::hasColumn('cpa', 'program_id'))
+                $table->dropColumn('program_id');
         });
     }
 }
